+++ mobile/src/main/java/com/alexstyl/specialdates/addevent/AvatarView.java
@@ -1,0 +1,6 @@
+package com.alexstyl.specialdates.addevent;
+
+import com.nostra13.universalimageloader.core.imageaware.ImageAware;
+
+interface AvatarView extends ImageAware {
+}
