+++ mobile/src/main/java/com/alexstyl/specialdates/addevent/ToolbarBackgroundStubAnimator.java
@@ -1,0 +1,14 @@
+package com.alexstyl.specialdates.addevent;
+
+final class ToolbarBackgroundStubAnimator implements ToolbarBackgroundAnimator {
+
+    @Override
+    public void fadeOut() {
+        // does nothing
+    }
+
+    @Override
+    public void fadeIn() {
+        // does nothing
+    }
+}
