+++ mobile/src/main/res/values-land/bools.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+  <bool name="isLandscape">true</bool>
+</resources>
